Update tests/test-support to use new KMP features (#18)
In our upgraded kotlin, the coroutines test module now has proper KMP
support. This PR updates our (currently private) test support module to
use the new `runTest` method, and updates our other methods / frameworks
accordingly to support it.

The big change here is that in the new framework we must create a
special coroutine for our `StoreFlow`s that doesn't get cancelled until
after the test has run. To help manage this we created the
`StoreManager` and `runStoreTest` methods. We'll likely need to ship a
public test-support module to account for this so keep an eye out for
this in the future.